+++ pkg/archive/archive_test.go
@@ -132,3 +132,114 @@ func TestTarGzDir(t *testing.T) {
 		t.Errorf("expected '%q' but got '%q'", string(original2), string(copied2))
 	}
 }
+
+func TestTarDDC(t *testing.T) {
+	src := filepath.Join("testdata", "ddctgz")
+	tmpDir := t.TempDir()
+	dest := filepath.Join(tmpDir, "output.tgz")
+	if err := archive.TarDDC(src, dest, "2050101011-DDC"); err != nil {
+		t.Fatalf("unable to archive file due to error %v", err)
+	}
+	f, err := os.Open(dest)
+	if err != nil {
+		t.Fatalf("unable to continue due to error %v", err)
+	}
+	zr, err := gzip.NewReader(f)
+	if err != nil {
+		t.Fatal(err)
+	}
+	outf, err := os.Create(filepath.Join(tmpDir, "output.tar"))
+	if err != nil {
+		t.Fatalf("unable to open file for writing with error %v", err)
+	}
+	if _, err = io.Copy(outf, zr); err != nil {
+		t.Fatalf("unable to copy file out %v", err)
+	}
+
+	if err := zr.Close(); err != nil {
+		t.Fatalf("unable to close zip read %v", err)
+	}
+
+	if err := f.Close(); err != nil {
+		t.Fatalf("unable to close destination tar %v", err)
+	}
+	if err := outf.Close(); err != nil {
+		t.Fatalf("unable to close output tar %v", err)
+	}
+
+	tarFile, err := os.Open(filepath.Join(tmpDir, "output.tar"))
+	if err != nil {
+		t.Fatalf("unable to read output tar file")
+	}
+	// Open and iterate through the files in the archive.
+	tr := tar.NewReader(tarFile)
+	for {
+		hdr, err := tr.Next()
+		if err == io.EOF {
+			break // End of archive
+		}
+		if err != nil {
+			t.Fatal(err)
+		}
+		outPath := filepath.Join(tmpDir, hdr.Name)
+
+		if hdr.Typeflag == tar.TypeDir {
+			if err := os.MkdirAll(outPath, 0700); err != nil {
+				t.Fatalf("unable to create dir path %v: %v", outPath, err)
+			}
+			continue
+		}
+		f, err := os.Create(outPath)
+		if err != nil {
+			t.Fatalf("unable to create path %v: %v", outPath, err)
+		}
+		if _, err := io.Copy(f, tr); err != nil {
+			t.Fatalf("unable to copy file %v out: %v", hdr.Name, err)
+		}
+		if err := f.Close(); err != nil {
+			t.Fatalf("unable to close file %v", err)
+		}
+	}
+	if err := tarFile.Close(); err != nil {
+		t.Fatal(err)
+	}
+	entries, err := os.ReadDir(tmpDir)
+	if err != nil {
+		t.Fatalf("unable to read test dir for logging %v", err)
+	}
+	for i, e := range entries {
+		t.Logf("entry #%v - %v", i, e)
+	}
+	_, err = os.Stat(filepath.Join(tmpDir, "2050101011-DDC", "file1.txt"))
+	if err != nil {
+		t.Fatalf("file missing due to error %v", err)
+	}
+
+	_, err = os.Stat(filepath.Join(tmpDir, "2050101011-DDC", "file2.txt"))
+	if err != nil {
+		t.Fatalf("file missing due to error %v", err)
+	}
+
+	copied1, err := os.ReadFile(filepath.Join(tmpDir, "2050101011-DDC", "file1.txt"))
+	if err != nil {
+		t.Fatalf("not able to read coped file1.txt: %v", err)
+	}
+	original1, err := os.ReadFile(filepath.Join("testdata", "ddctgz", "2050101011-DDC", "file1.txt"))
+	if err != nil {
+		t.Fatalf("unable to read origina file1.txt file: %v", err)
+	}
+	if !reflect.DeepEqual(copied1, original1) {
+		t.Errorf("expected '%q' but got '%q'", string(original1), string(copied1))
+	}
+	copied2, err := os.ReadFile(filepath.Join(tmpDir, "2050101011-DDC", "file2.txt"))
+	if err != nil {
+		t.Fatalf("not able to read coped file2.txt: %v", err)
+	}
+	original2, err := os.ReadFile(filepath.Join("testdata", "ddctgz", "2050101011-DDC", "file2.txt"))
+	if err != nil {
+		t.Fatalf("unable to read original file2.txt file: %v", err)
+	}
+	if !reflect.DeepEqual(copied2, original2) {
+		t.Errorf("expected '%q' but got '%q'", string(original2), string(copied2))
+	}
+}